next work fps drop issue, thanks @Fewizz for rewriting grave flower render & AxisAlignedBB fix & Enum many data in heap fix; replace switch->if else if; new fake blocks, enable replaceableOreGen, add/change tickRate's; add GeneratorEntropy class - more entropy; fix(?) missing commads if enter more 1 world/world created; spawn mobs no need player, without boss spawner; add shift fog time, add chance fog work if raining; reduce calls getBlock, change spawn in onUpdate(); reduce value block in creeper room & change proportion & chance gen chest in catacombs; fix(?) infinity loop in gen component catacombs Fence; LinkedList/ArrayList->ArrayList/(array+cycle or EnumSet), less memory use/more fast; reduce grave & invisiblewall explosion resistance, add chance spawn mob if explosion grave; fix render NPE-crash if non-mod or non-vanilla sword grave, example Thaumcraft swords; +1 light level in GSMobSpawner MAX_LIGHT_VALUE; Override TileEnity markDirty() - to an empty function; fix infinity spawn mobs when open HauntedChest, more realism;

+++ src/main/java/gravestone/gui/GSGuiGrave.java
@@ -58,7 +58,7 @@ public class GSGuiGrave extends GuiScreen {
                 mc.displayGuiScreen((GuiScreen) null);
                 break;
         }
-        GSMessageHandler.networkWrapper.sendToServer(new GraveDeathMessageToServer(teGrave.getWorldObj(), teGrave.xCoord, teGrave.yCoord, teGrave.zCoord, this.textField.getText(), isRandomTextButtonClicked));
+        GSMessageHandler.networkWrapper.sendToServer(new GraveDeathMessageToServer(teGrave.getWorldObj(), teGrave.xCoord, (short)teGrave.yCoord, teGrave.zCoord, this.textField.getText(), isRandomTextButtonClicked));
     }
 
     @Override
